refactor: organized src/ast
