+++ src/ast/expression.rs
@@ -1,57 +1,6 @@
 use crate::lexer::Token;
 
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Identifier {
-    pub token: Token,
-    pub value: String,
-}
-
-impl std::fmt::Display for Identifier {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}", &self.value)
-    }
-}
-
-#[derive(Clone, PartialEq, Eq)]
-pub enum Literal {
-    Integer(i64),
-    Boolean(bool),
-    String(String),
-    Array(Vec<Expression>),
-    Hash(Vec<(Expression, Expression)>),
-}
-
-impl std::fmt::Display for Literal {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Literal::Integer(int) => write!(f, "{}", int),
-            Literal::Boolean(bool) => write!(f, "{}", bool),
-            Literal::String(string) => write!(f, "{}", string),
-            Literal::Array(array) => {
-                write!(f, "[")?;
-                for (i, expr) in array.iter().enumerate() {
-                    if i == array.len() - 1 {
-                        write!(f, "{}", expr)?;
-                    } else {
-                        write!(f, "{}, ", expr)?;
-                    }
-                }
-                write!(f, "]")
-            }
-            Literal::Hash(hash) => {
-                write!(f, "{{")?;
-                for (i, (key, value)) in hash.iter().enumerate() {
-                    if i == hash.len() - 1 {
-                        write!(f, "{}: {}", key, value)?;
-                    } else {
-                        write!(f, "{}: {}, ", key, value)?;
-                    }
-                }
-                write!(f, "}}")
-            }
-        }
-    }
-}
+use super::{literal::Literal, statement::BlockStatement, Identifier};
 
 #[derive(Clone, PartialEq, Eq)]
 pub enum Expression {
@@ -155,43 +104,3 @@ impl std::fmt::Display for Expression {
         }
     }
 }
-
-#[derive(Clone, PartialEq, Eq)]
-pub enum Statement {
-    Let {
-        token: Token,
-        name: Identifier,
-        value: Expression,
-    },
-    ReAssign {
-        token: Token,
-        name: Identifier,
-        value: Expression,
-    },
-    Return {
-        token: Token,
-        value: Expression,
-    },
-    Expression {
-        token: Token,
-        value: Expression,
-    },
-}
-
-impl std::fmt::Debug for Statement {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Statement::Let { name, value, .. } => {
-                write!(f, "let {} = {}) ", name, value)
-            }
-            Statement::Return { value, .. } => write!(f, "return {};", value),
-            Statement::Expression { value, .. } => write!(f, "{}", value),
-            Statement::ReAssign { name, value, .. } => {
-                write!(f, "reassign {} = {}", name, value)
-            }
-        }
-    }
-}
-
-pub type BlockStatement = Vec<Statement>;
-pub type Program = BlockStatement;
